+++ biodumpy/inputs/IUCN.py
@@ -5,50 +5,57 @@ from biodumpy import Input, BiodumpyException
 
 class IUCN(Input):
 	"""
-	Query the IUCN Red List API for information about a specific species in a given region.
+    Query the IUCN Red List API for information about a specific species in a given region.
 
-	Parameters
-	----------
-	query : list
-	    The list of taxa to query.
-	api_key : str
-	    Your IUCN API key.
-	habitat : str
-	    The name of an IUCN habitat.
-	historical : bool, optional
-	    If True, the function returns also the historical threats of a taxon. Default is False.
-	threats : bool, optional
-	    If True, the function returns also the threat list of a taxon. Default is False.
-	regions : list, optional
-	    The list of specific IUCN regions. For further information, see the IUCN API documentation. Default is an empty list.
-	bulk : bool, optional
-		If True, the function creates a bulk file. For further information, see the documentation of the Biodumpy package. Default is False.
-	output_format : string, optional
-		The format of the output file. The options available are: 'json', 'fasta', 'pdf'. Default is 'json'.
+    Parameters
+    ----------
+    query : list
+        The list of taxa to query.
+    api_key : str
+        Your IUCN API key.
+    habitat : str
+        The name of an IUCN habitat.
+    regions : list, optional
+        The list of specific IUCN regions. For further information, see the IUCN API documentation. Default is an empty list.
+    historical : bool, optional
+        If True, the function returns also the historical threats of a taxon. Default is False.
+    threats : bool, optional
+        If True, the function returns also the threat list of a taxon. Default is False.
+    citation : bool, optional
+        If True, the function returns the citation for a given species assessment.
+    weblink : bool, optional
+        If True, the function returns the link to redirect to the Red List website page for this species.
+    bulk : bool, optional
+        If True, the function creates a bulk file. For further information, see the documentation of the Biodumpy package.
+        Default is False.
+    output_format : string, optional
+        The format of the output file. The options available are: 'json', 'fasta', 'pdf'. Default is 'json'.
 
-	Example
-	-------
-	>>> from biodumpy import Biodumpy
-	>>> from biodumpy.inputs import IUCN
-	# Insert your API key
-	>>> api_key = 'YOUR_API_KEY'
-	# Taxa list
-	>>> taxa = ['Alytes muletensis', 'Bufotes viridis', 'Hyla meridionalis']
-	# Select your output path
-	>>> output_path = 'YOUR_OUTPUT_PATH'
-	# Select the regions
-	>>> regions = ['global', 'europe']
-	>>> bdp = Biodumpy([IUCN(api_key=api_key, bulk=True, regions=regions)])
-	>>> bdp.start(taxa, output_path=output_path)
+    Example
+    -------
+    >>> from biodumpy import Biodumpy
+    >>> from biodumpy.inputs import IUCN
+    # Insert your API key
+    >>> api_key = 'YOUR_API_KEY'
+    # Taxa list
+    >>> taxa = ['Alytes muletensis', 'Bufotes viridis', 'Hyla meridionalis']
+    # Select your output path
+    >>> output_path = 'YOUR_OUTPUT_PATH'
+    # Select the regions
+    >>> regions = ['global', 'europe']
+    >>> bdp = Biodumpy([IUCN(api_key=api_key, bulk=True, regions=regions)])
+    >>> bdp.start(taxa, output_path=output_path)
 	"""
 
 	def __init__(
 		self,
 		api_key: str,
 		habitat: bool = False,
+		regions: list = None,
 		historical: bool = False,
 		threats: bool = False,
-		regions: list = None,
+		citation: bool = False,
+		weblink: bool = False,
 		output_format: str = "json",
 		bulk: bool = False,
 	):
@@ -60,6 +67,8 @@ class IUCN(Input):
 		self.regions = regions
 		self.historical = historical
 		self.threats = threats
+		self.citation = citation
+		self.weblink = weblink
 
 		iucn_regions = [
 			"northern_africa",
@@ -71,7 +80,7 @@ class IUCN(Input):
 			"western_africa",
 			"southern_africa",
 			"mediterranean",
-			"europe",
+			"europe"
 		]
 
 		if output_format != "json":
@@ -85,11 +94,13 @@ class IUCN(Input):
 		payload = []
 
 		for region in self.regions:
-			taxon_info = self._icun_request(f"https://apiv3.iucnredlist.org/api/v3/species/{query}/region/{region}?token={self.api_key}")
+			taxon_info = self._icun_request(
+				f"https://apiv3.iucnredlist.org/api/v3/species/{query}/region/{region}?token={self.api_key}")
 
-			if not taxon_info or taxon_info.get("taxonid") is None:
+			if not taxon_info or taxon_info[0].get("taxonid") is None:
 				continue
 
+			taxon_info = taxon_info[0]
 			payload.append(taxon_info)
 			taxon_id = taxon_info["taxonid"]
 
@@ -101,16 +112,26 @@ class IUCN(Input):
 					taxon_info.update({"habitat": habitat_info})
 
 				if self.historical:
-					habitat_info = self._icun_request(
+					historical_info = self._icun_request(
 						f"https://apiv3.iucnredlist.org/api/v3/species/history/id/{taxon_id}/region/{region}?token={self.api_key}"
 					)
-					taxon_info.update({"historical": habitat_info})
+					taxon_info.update({"historical": historical_info})
 
 				if self.threats:
-					habitat_info = self._icun_request(
+					threats_info = self._icun_request(
 						f"https://apiv3.iucnredlist.org/api/v3/threats/species/id/{taxon_id}/region/{region}?token={self.api_key}"
 					)
-					taxon_info.update({"threats": habitat_info})
+					taxon_info.update({"threats": threats_info})
+
+				if self.citation:
+					citation_info = self._icun_request(
+						f"https://apiv3.iucnredlist.org/api/v3/species/citation/id/{taxon_id}/region/{region}?token={self.api_key}"
+					)
+					taxon_info.update({"citation": citation_info})
+
+				if self.weblink:
+					weblink_info = f"https://apiv3.iucnredlist.org/api/v3/taxonredirect/{taxon_id}/{region}"
+					taxon_info.update({"weblink": weblink_info})
 
 		return payload
 
@@ -125,7 +146,7 @@ class IUCN(Input):
 			dict: A dictionary containing information about the species in the specified region if found.
 
 		Example:
-			result = taxon_iucn(f"https://apiv3.iucnredlist.org/api/v3/habitats/species/id/{taxon_id}/region/{region}?token={self.api_key}")
+			result = icun_request(f"https://apiv3.iucnredlist.org/api/v3/habitats/species/id/{taxon_id}/region/{region}?token={self.api_key}")
 		"""
 		response = requests.get(query_path)
 
@@ -136,8 +157,9 @@ class IUCN(Input):
 			response = response.json()
 
 			if not (len(response.get("result", [])) == 0 or (response.get("value") == "0")):
-				result = response.get("result", [])[0]
+				result = response.get("result", [])
 				if "region_identifier" in response:
-					result.update({"region": response["region_identifier"]})
+					for r in result:
+						r.update({"region": response["region_identifier"]})
 
 				return result