+++ biodumpy/inputs/__init__.py
@@ -7,3 +7,5 @@ from .ZooBank import ZooBank
 from .IUCN import IUCN
 from .GBIF import GBIF
 from .OBIS import OBIS
+from .Crossref import Crossref
+from .WORMS import WORMS