+++ tests/IUCN.py
@@ -12,7 +12,8 @@ from biodumpy.inputs import IUCN
 # set a trap and redirect stdout. Remove the print of the function. In this wat the test output is cleanest.
 trap = io.StringIO()
 
-API_KEY = "YOUR_API_KEY"
+
+API_KEY = None
 
 IUCN_REGIONS = [
 	"northern_africa",
@@ -24,11 +25,11 @@ IUCN_REGIONS = [
 	"western_africa",
 	"southern_africa",
 	"mediterranean",
-	"europe",
+	"europe"
 ]
 
 
-def iucn_query(query, api_key, habitat, regions, historical, threats, output_format):
+def iucn_query(query, api_key, habitat, regions, historical, threats, weblink, output_format):
 	# Create temporary directory
 	with tempfile.TemporaryDirectory() as temp_dir:
 		# Construct the dynamic path using formatted strings
@@ -44,7 +45,8 @@ def iucn_query(query, api_key, habitat, regions, historical, threats, output_for
 				regions=regions,
 				historical=historical,
 				threats=threats,
-				output_format=output_format,
+				weblink=weblink,
+				output_format=output_format
 			)
 		]
 	)
@@ -71,6 +73,7 @@ def test_iucn_initialization():
 	assert iucn.habitat == False
 	assert iucn.historical == False
 	assert iucn.threats == False
+	assert iucn.weblink == False
 	assert iucn.output_format == "json"
 
 	# Objective: Verify that the class raises a ValueError when an invalid value is provided for the
@@ -103,15 +106,16 @@ def test_validate_regions_invalid():
 
 
 @pytest.mark.parametrize(
-	"query, regions, habitat, historical, threats, output_format",
+	"query, regions, habitat, historical, threats, weblink, output_format",
 	[
-		(["Alytes muletensis"], ["global"], False, False, False, "json"),
-		(["Alytes muletensis"], ["global"], True, False, False, "json"),
-		(["Alytes muletensis"], ["global"], False, True, False, "json"),
-		(["Alytes muletensis"], ["global"], False, False, True, "json"),
-	],
+		(["Alytes muletensis"], ["global"], False, False, False, False, "json"),
+		(["Alytes muletensis"], ["global"], True, False, False, False, "json"),
+		(["Alytes muletensis"], ["global"], False, True, False, False, "json"),
+		(["Alytes muletensis"], ["global"], False, False, True, False, "json"),
+		(["Alytes muletensis"], ["global"], False, False, False, True, "json")
+	]
 )
-def test_download(query, regions, habitat, historical, threats, output_format):
+def test_download(query, regions, habitat, historical, threats, weblink, output_format):
 	with redirect_stdout(trap):
 		data = iucn_query(
 			query=query,
@@ -120,6 +124,7 @@ def test_download(query, regions, habitat, historical, threats, output_format):
 			historical=historical,
 			threats=threats,
 			output_format=output_format,
+			weblink=weblink,
 			api_key=API_KEY,
 		)
 
@@ -129,31 +134,93 @@ def test_download(query, regions, habitat, historical, threats, output_format):
 	# Check some fields
 	data = data[0]
 	assert "taxonid" in data, "taxonid is not in data"
+	assert data["taxonid"] == 977, "the taxonid is not 977"
+
 	assert "scientific_name" in data, "scientific_name is not in data"
+	assert data["scientific_name"] == 'Alytes muletensis', "the scientific_name is not Alytes muletensis"
+
 	assert "category" in data, "category is not in data"
+	assert data["category"] == 'EN', "the category is not EN"
+
 	assert "assessment_date" in data, "assessment_date is not in data"
+	assert data["assessment_date"] == '2020-04-23', "the assessment_date is not 2020-04-23"
+
 	assert "region" in data, "region is not in data"
-	assert data["region"] == "global", "assessment_date is not in data"
+	assert data["region"] == "global", "the region is not global"
 
 	if habitat:
 		assert "habitat" in data, "habitat is not in data"
-		assert "code" in data["habitat"], "code is not in data['habitat']"
-		assert "habitat" in data["habitat"], "habitat is not in data['habitat']"
-		assert "suitability" in data["habitat"], "suitability is not in data['habitat']"
-		assert "season" in data["habitat"], "season is not in data['habitat']"
+
+		habitat = data["habitat"]
+
+		assert len(habitat) == 4, "the number of habitat is 4"
+
+		assert "code" in habitat[0], "code is not in habitat[0]"
+		assert habitat[0]["code"] == "14.2", "the code is not 14.2"
+
+		assert "habitat" in habitat[0], "habitat is not in habitat[0]"
+		assert habitat[0][
+			       "habitat"] == 'Artificial/Terrestrial - Pastureland', "the habitat is not Artificial/Terrestrial - Pastureland"
+
+		assert "suitability" in habitat[0], "suitability is not in habitat[0]"
+		assert habitat[0]['suitability'] == 'Suitable', "the suitability is not Suitable"
+
+		assert "season" in habitat[0], "season is not in habitat[0]"
+		assert habitat[0]['season'] == 'Resident', "the season is not Resident"
+
+		assert "majorimportance" in habitat[0], "majorimportance is not in habitat[0]"
+		assert habitat[0]['majorimportance'] is None, "the majorimportance is not None"
 
 	if historical:
 		assert "historical" in data, "historical is not in data"
-		assert "year" in data["historical"], "code is not in data['historical']"
-		assert "assess_year" in data["historical"], "assess_year is not in data['historical']"
-		assert "code" in data["historical"], "code is not in data['historical']"
-		assert "category" in data["historical"], "category is not in data['historical']"
-		assert "region" in data["historical"], "region is not in data['historical']"
+
+		historical = data["historical"]
+
+		assert len(historical) == 7, "the number of historical records is 7"
+
+		assert "year" in historical[0], "code is not in historical[0]"
+		assert historical[0]['year'] == "2020", "the year is not 2020"
+
+		assert "assess_year" in historical[0], "assess_year is not in historical[0]"
+		assert historical[0]['assess_year'] == "2020", "the year is not 2020"
+
+		assert "code" in historical[0], "code is not in historical[0]"
+		assert historical[0]['code'] == "EN", "the code is not EN"
+
+		assert "category" in historical[0], "category is not in historical[0]"
+		assert historical[0]['category'] == "Endangered", "the category is not Endangered"
+
+		assert "region" in historical[0], "region is not in historical[0]"
+		assert historical[0]['region'] == "global", "the region is not global"
 
 	if threats:
 		assert "threats" in data, "threats is not in data"
-		assert "title" in data["threats"], "title is not in data['threats']"
-		assert "timing" in data["threats"], "timing is not in data['threats']"
-		assert "code" in data["threats"], "code is not in data['threats']"
-		assert "scope" in data["threats"], "scope is not in data['threats']"
-		assert "score" in data["threats"], "score is not in data['threats']"
+
+		threats = data["threats"]
+
+		assert len(threats) == 12, "the number of threats records is 12"
+
+		assert "code" in threats[0], "code is not in threats[0]"
+		assert threats[0]['code'] == "1.1", "the code is not 1.1"
+
+		assert "title" in threats[0], "title is not in threats[0]"
+		assert threats[0]['title'] == "Housing & urban areas", "the title is not Housing & urban areas"
+
+		assert "timing" in threats[0], "timing is not in threats[0]"
+		assert threats[0]['timing'] == "Ongoing", "the timing is not Ongoing"
+
+		assert "scope" in threats[0], "scope is not in threats[0]"
+		assert threats[0]['scope'] is None, "the scope is not None"
+
+		assert "severity" in threats[0], "severity is not in threats[0]"
+		assert threats[0]['severity'] is None, "the severity is not None"
+
+		assert "score" in threats[0], "score is not in threats[0]"
+		assert threats[0]['score'] == 'Low Impact: 3', "the score is not Low Impact: 3"
+
+		assert "invasive" in threats[0], "invasive is not in threats[0]"
+		assert threats[0]['invasive'] is None, "the invasive is not None"
+
+	if weblink:
+		assert "weblink" in data, "weblink is not in data"
+		assert data["weblink"] == 'https://apiv3.iucnredlist.org/api/v3/taxonredirect/977/global', "The weblink is not correct"