+++ tests/NCBI.py
@@ -13,7 +13,8 @@ from biodumpy.inputs import NCBI
 trap = io.StringIO()
 
 
-def ncbi_query(query, summary, output_format, max_bp, db, step, rettype, query_type, by_id, mail):
+def ncbi_query(query, summary, output_format, max_bp, db, step, rettype,
+               query_type, by_id, taxonomy, taxonomy_only, mail):
 	# Create temporary directory
 	with tempfile.TemporaryDirectory() as temp_dir:
 		# Construct the dynamic path using formatted strings
@@ -32,7 +33,9 @@ def ncbi_query(query, summary, output_format, max_bp, db, step, rettype, query_t
 				rettype=rettype,
 				query_type=query_type,
 				by_id=by_id,
-				mail=mail,
+				taxonomy=taxonomy,
+				taxonomy_only=taxonomy_only,
+				mail=mail
 			)
 		]
 	)
@@ -66,6 +69,8 @@ def test_ncbi_initialization():
 	assert ncbi.max_bp is None
 	assert ncbi.summary == False
 	assert ncbi.by_id == False
+	assert ncbi.taxonomy == False
+	assert ncbi.taxonomy_only == False
 	assert ncbi.bulk == False
 	assert ncbi.output_format == "json"
 
@@ -78,21 +83,28 @@ def test_ncbi_initialization():
 	with pytest.raises(ValueError, match="Invalid parameters: 'summary' is True, so 'output_format' cannot be 'fasta'."):
 		NCBI(summary=True, output_format="fasta", rettype="fasta")  # Should raise the error
 
+	with pytest.raises(ValueError, match="Invalid parameters: 'taxonomy' is True, so 'output_format' cannot be 'fasta'."):
+		NCBI(taxonomy=True, output_format="fasta", rettype="fasta")  # Should raise the error
+
+	with pytest.raises(ValueError, match="Invalid parameters: 'taxonomy_only' is True, so 'output_format' cannot be 'fasta'."):
+		NCBI(taxonomy_only=True, output_format="fasta", rettype="fasta")  # Should raise the error
+
 	with pytest.raises(ValueError, match='Invalid output_format. Expected "json" or "fasta".'):
 		NCBI(output_format="xml")  # Should raise the error
 
 
 @pytest.mark.parametrize(
-	"query, summary, output_format, max_bp, db, step, rettype, query_type, by_id, mail",
+	"query, summary, output_format, max_bp, db, step, rettype, query_type, by_id, taxonomy, taxonomy_only, mail",
 	[
-		(["Anax imperator"], False, "json", 5000, "nucleotide", 100, "gb", "[Organism]", False, "hola@quetal.com"),
-		(["Anax imperator"], False, "json", 5000, "nucleotide", 100, "gb", "[Organism] AND COX1[Gene]", False, "hola@quetal.com"),
-		(["OQ507551"], False, "json", 5000, "nucleotide", 100, "gb", None, True, "hola@quetal.com"),
-		(["Anax imperator"], True, "json", 5000, "nucleotide", 100, "gb", "[Organism]", False, "hola@quetal.com"),
-		(["Anax imperator"], False, "fasta", 5000, "nucleotide", 100, "fasta", "[Organism]", False, "hola@quetal.com"),
-	],
+		(["Alytes muletensis"], False, "json", 2000, "nucleotide", 100, "gb", "[Organism]", False, False, False, "hola@quetal.com"),
+		(["Alytes muletensis"], False, "json", 2000, "nucleotide", 100, "gb", "[Organism] AND COX1[Gene]", False, False, False, "hola@quetal.com"),
+		(["AY166960"], False, "json", 2000, "nucleotide", 100, "gb", None, True, False, False, "hola@quetal.com"),
+		(["Alytes muletensis"], True, "json", 2000, "nucleotide", 100, "gb", "[Organism]", False, False, False, "hola@quetal.com"),
+		(["Alytes muletensis"], False, "fasta", 2000, "nucleotide", 100, "fasta", "[Organism]", False, False, False, "hola@quetal.com"),
+	]
 )
-def test_download(query, summary, output_format, max_bp, db, step, rettype, query_type, by_id, mail):
+def test_download(query, summary, output_format, max_bp, db, step, rettype,
+                  query_type, by_id, taxonomy, taxonomy_only, mail):
 	with redirect_stdout(trap):
 		data = ncbi_query(
 			query=query,
@@ -104,7 +116,9 @@ def test_download(query, summary, output_format, max_bp, db, step, rettype, quer
 			rettype=rettype,
 			query_type=query_type,
 			by_id=by_id,
-			mail=mail,
+			taxonomy=taxonomy,
+			taxonomy_only=taxonomy_only,
+			mail=mail
 		)
 
 	# Check if data is not empty
@@ -116,7 +130,16 @@ def test_download(query, summary, output_format, max_bp, db, step, rettype, quer
 		assert "id" in data, "id is not in data"
 		assert "name" in data, "name is not in data"
 		assert "description" in data, "description is not in data"
+
 		assert "annotations" in data, "annotations is not in data"
+		elem_annotations = ['molecule_type', 'topology', 'data_file_division', 'date', 'accessions', 'sequence_version', 'keywords', 'source', 'organism', 'taxonomy', 'references']
+		found_words = []
+		for word in elem_annotations:
+			if word in data["annotations"]:
+				found_words.append(word)
+		assert len(found_words) == 11, "Check the annotation list"
+
+
 		assert "features" in data, "features is not in data"
 
 	if summary and output_format != "fasta":
@@ -127,6 +150,58 @@ def test_download(query, summary, output_format, max_bp, db, step, rettype, quer
 		assert "Length" in data, "Length is not in data"
 		assert "query" in data, "query is not in data"
 
+		# Check if the length is lover than 2000
+		assert data["Length"] < 2000, "Length is not in correct, is higher than 2000 bp"
+
 	if summary is False and output_format == "fasta":
 		# Check if the fasta file starts with >
 		assert data.startswith(">")
+
+
+@pytest.mark.parametrize(
+	"query, summary, output_format, max_bp, db, step, rettype, query_type, by_id, taxonomy, taxonomy_only, mail",
+	[
+		(["Alytes muletensis"], False, "json", 2000, "nucleotide", 100, "gb", "[Organism]", False, True, False, "hola@quetal.com"),
+		(["Alytes muletensis"], False, "json", 2000, "nucleotide", 100, "gb", "[Organism]", False, False, True, "hola@quetal.com")
+	]
+)
+def test_download_taxonomy(query, summary, output_format, max_bp, db, step, rettype,
+                           query_type, by_id, taxonomy, taxonomy_only, mail):
+	with redirect_stdout(trap):
+		data = ncbi_query(
+			query=query,
+			summary=summary,
+			output_format=output_format,
+			max_bp=max_bp,
+			db=db,
+			step=step,
+			rettype=rettype,
+			query_type=query_type,
+			by_id=by_id,
+			taxonomy=taxonomy,
+			taxonomy_only=taxonomy_only,
+			mail=mail
+		)
+
+	# Check if data is not empty
+	assert len(data) > 0, "data length is 0"
+
+	if taxonomy:
+		data = data[-1]
+		assert "taxonomy" in data, "taxonomy is not in data"
+
+		assert len(data["taxonomy"]) == 23, "The length of taxonomy is not 23"
+
+		data = data["taxonomy"]
+		assert 'TaxId' in data[0], "TaxId is not in data[0]"
+		assert 'ScientificName' in data[0], "ScientificName is not in data[0]"
+		assert 'Rank' in data[0], "Rank is not in data[0]"
+
+	if taxonomy is False and taxonomy_only:
+		data = data[0]
+
+		assert len(data) == 23, "The length of taxonomy is not 23"
+
+		assert 'TaxId' in data[0], "TaxId is not in data[0]"
+		assert 'ScientificName' in data[0], "ScientificName is not in data[0]"
+		assert 'Rank' in data[0], "Rank is not in data[0]"