+++ CHANGELOG.md
@@ -1,14 +1,30 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+All notable changes to this project will be documented in this file 🐔.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+---
+
+## [0.1.4] - Unreleased
 
 ### Added
-- Initial release of **biodumpy** with support to download metadata for BOLD, COL, GBIF, iNaturalist, IUCN, NCBI, OBIS,
-ZooBank
+- Crossref module.
+- WORMS module.
+- Parameter weblink in the IUCN module.
+- Parameters taxonomy and taxonomy_only in the NCBI module.
+- Sections "The adaptation of biodumpy to R" and "Bibliography" in the documentation.
+
+### Changed
+- Improved documentation.
+- Enhanced tests for the IUCN and NCBI modules.
+
+---
+
+## [0.1.3] - 2024-09-20
+
+### Added
+- Initial release of biodumpy with support for downloading metadata from BOLD, COL, GBIF, iNaturalist, IUCN, NCBI, OBIS, and ZooBank.
 
 ### Changed
 - Improved documentation.
@@ -16,6 +32,13 @@ ZooBank
 ### Fixed
 - Bug fixes in the download modules.
 
+---
+
 ## [0.1.2] - 2024-09-20
+
 ### Added
 - Initial release for testing.
+
+---
+
+## [Unreleased]